# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/minerals": {
            "get": {
                "description": "Возвращает список самоцветов с поддержкой фильтрации, сортировки и пагинации",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "minerals"
                ],
                "summary": "Получить список минералов",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Язык ответа",
                        "name": "lang",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Режим отображения",
                        "name": "view",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Редкость",
                        "name": "rarity",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Группа минерала",
                        "name": "mineral_group",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Цвет",
                        "name": "color",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Только российские",
                        "name": "russian_only",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Минимальная твёрдость",
                        "name": "hardness_min",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Максимальная твёрдость",
                        "name": "hardness_max",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Сортировка",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Порядок сортировки",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Количество на странице",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Страница",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ListResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Создаёт новый самоцвет/минерал в базе (админка)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "minerals"
                ],
                "summary": "Создать новый минерал",
                "parameters": [
                    {
                        "description": "Данные минерала",
                        "name": "mineral",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.CreateMineralRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_roslava_samotsvety-api_internal_domain.Mineral"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Slug уже существует",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/minerals/{slug}": {
            "get": {
                "description": "Возвращает полную детальную карточку самоцвета",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "minerals"
                ],
                "summary": "Получить минерал по slug",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Slug минерала",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Язык",
                        "name": "lang",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Режим",
                        "name": "view",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_roslava_samotsvety-api_internal_domain.Mineral"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет минерал по slug (админка)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "minerals"
                ],
                "summary": "Удалить минерал",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Slug минерала",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/search": {
            "get": {
                "description": "Полнотекстовый поиск по названию, синонимам, lore и химической формуле (работает на ru+en)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "search"
                ],
                "summary": "Поиск минералов",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Поисковый запрос",
                        "name": "q",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Язык ответа",
                        "name": "lang",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Режим",
                        "name": "view",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Лимит",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Страница",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ListResponse"
                        }
                    }
                }
            }
        },
        "/minerals/{slug}": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "minerals"
                ],
                "summary": "Обновить минерал",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Slug минерала",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные для обновления",
                        "name": "mineral",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.UpdateMineralRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_roslava_samotsvety-api_internal_domain.Mineral"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_roslava_samotsvety-api_internal_domain.GalleryImage": {
            "type": "object",
            "properties": {
                "description_en": {
                    "type": "string"
                },
                "description_ru": {
                    "type": "string"
                },
                "type": {
                    "description": "specimen, polished, jewelry, micro и т.д.",
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "github_com_roslava_samotsvety-api_internal_domain.Hardness": {
            "type": "object",
            "required": [
                "max",
                "min"
            ],
            "properties": {
                "max": {
                    "type": "number",
                    "maximum": 10,
                    "minimum": 1
                },
                "min": {
                    "type": "number",
                    "maximum": 10,
                    "minimum": 1
                },
                "note": {
                    "type": "string"
                }
            }
        },
        "github_com_roslava_samotsvety-api_internal_domain.I18n": {
            "type": "object",
            "properties": {
                "en": {
                    "$ref": "#/definitions/github_com_roslava_samotsvety-api_internal_domain.LangData"
                },
                "ru": {
                    "$ref": "#/definitions/github_com_roslava_samotsvety-api_internal_domain.LangData"
                }
            }
        },
        "github_com_roslava_samotsvety-api_internal_domain.LangData": {
            "type": "object",
            "required": [
                "lore",
                "name"
            ],
            "properties": {
                "color": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "color_description": {
                    "type": "string"
                },
                "esoteric": {
                    "$ref": "#/definitions/github_com_roslava_samotsvety-api_internal_domain.Esoteric"
                },
                "lore": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "synonyms": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "github_com_roslava_samotsvety-api_internal_domain.Locality": {
            "type": "object",
            "properties": {
                "country": {
                    "type": "string"
                },
                "description_en": {
                    "type": "string"
                },
                "description_ru": {
                    "type": "string"
                },
                "famous": {
                    "type": "boolean"
                },
                "is_russian": {
                    "type": "boolean"
                },
                "locality": {
                    "type": "string"
                },
                "region": {
                    "type": "string"
                }
            }
        },
        "github_com_roslava_samotsvety-api_internal_domain.Mineral": {
            "type": "object",
            "required": [
                "slug"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "gallery": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_roslava_samotsvety-api_internal_domain.GalleryImage"
                    }
                },
                "i18n": {
                    "$ref": "#/definitions/github_com_roslava_samotsvety-api_internal_domain.I18n"
                },
                "localities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_roslava_samotsvety-api_internal_domain.Locality"
                    }
                },
                "main_image_url": {
                    "type": "string"
                },
                "related_minerals": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "safety_notes": {
                    "type": "string"
                },
                "scientific": {
                    "$ref": "#/definitions/github_com_roslava_samotsvety-api_internal_domain.Scientific"
                },
                "slug": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_roslava_samotsvety-api_internal_domain.Rarity": {
            "type": "string",
            "enum": [
                "common",
                "uncommon",
                "rare",
                "very_rare"
            ],
            "x-enum-varnames": [
                "RarityCommon",
                "RarityUncommon",
                "RarityRare",
                "RarityVeryRare"
            ]
        },
        "github_com_roslava_samotsvety-api_internal_domain.Scientific": {
            "type": "object",
            "required": [
                "chemical_formula",
                "crystal_system",
                "hardness",
                "luster",
                "mineral_group",
                "rarity",
                "specific_gravity",
                "streak",
                "transparency"
            ],
            "properties": {
                "chemical_formula": {
                    "type": "string"
                },
                "cleavage": {
                    "type": "string"
                },
                "crystal_habit": {
                    "type": "string"
                },
                "crystal_system": {
                    "type": "string"
                },
                "fracture": {
                    "type": "string"
                },
                "hardness": {
                    "$ref": "#/definitions/github_com_roslava_samotsvety-api_internal_domain.Hardness"
                },
                "identification_tips": {
                    "type": "string"
                },
                "ima_status": {
                    "type": "string"
                },
                "luster": {
                    "type": "string"
                },
                "mineral_group": {
                    "type": "string"
                },
                "rarity": {
                    "enum": [
                        "common",
                        "uncommon",
                        "rare",
                        "very_rare"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_roslava_samotsvety-api_internal_domain.Rarity"
                        }
                    ]
                },
                "specific_gravity": {
                    "$ref": "#/definitions/github_com_roslava_samotsvety-api_internal_domain.SpecificGravity"
                },
                "streak": {
                    "type": "string"
                },
                "tenacity": {
                    "type": "string"
                },
                "transparency": {
                    "type": "string"
                }
            }
        },
        "github_com_roslava_samotsvety-api_internal_domain.SpecificGravity": {
            "type": "object",
            "required": [
                "max",
                "min"
            ],
            "properties": {
                "max": {
                    "type": "number",
                    "minimum": 1
                },
                "min": {
                    "type": "number",
                    "minimum": 1
                }
            }
        },
...
        "internal_handler.CreateMineralRequest": {
            "type": "object",
            "required": [
                "i18n",
                "scientific",
                "slug"
            ],
            "properties": {
                "gallery": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_roslava_samotsvety-api_internal_domain.GalleryImage"
                    }
                },
                "i18n": {
                    "$ref": "#/definitions/github_com_roslava_samotsvety-api_internal_domain.I18n"
                },
                "localities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_roslava_samotsvety-api_internal_domain.Locality"
                    }
                },
                "main_image_url": {
                    "type": "string"
                },
                "related_minerals": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "safety_notes": {
                    "type": "string"
                },
                "scientific": {
                    "$ref": "#/definitions/github_com_roslava_samotsvety-api_internal_domain.Scientific"
                },
                "slug": {
                    "type": "string"
                }
            }
        },
        "internal_handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "internal_handler.ListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_roslava_samotsvety-api_internal_domain.Mineral"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "internal_handler.UpdateMineralRequest": {
            "type": "object",
            "properties": {
                "gallery": {
                    "description": "указатель",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_roslava_samotsvety-api_internal_domain.GalleryImage"
                    }
                },
                "i18n": {
                    "$ref": "#/definitions/github_com_roslava_samotsvety-api_internal_domain.I18n"
                },
                "localities": {
                    "description": "указатель — важно!",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_roslava_samotsvety-api_internal_domain.Locality"
                    }
                },
                "main_image_url": {
                    "type": "string"
                },
                "related_minerals": {
                    "description": "указатель",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "safety_notes": {
                    "type": "string"
                },
                "scientific": {
                    "$ref": "#/definitions/github_com_roslava_samotsvety-api_internal_domain.Scientific"
                },
                "slug": {
                    "type": "string"
                }
            }
        }
    }
}